refactor: 로컬 스토리지 관련 상수 분리
